feat(api): JSON GET / probe for direct uvicorn; legacy BMI HTML path
- GET / returns stable application/json probe (service, surface, links)
- Embedded legacy BMI calculator HTML at GET /legacy/bmi-calculator
- Caddy: proxy /legacy* to API to avoid SPA try_files swallow
- Contract + README + deploy AGENTS; ledger item closed
- Tests: contract probe, retarget HTML expectations, Spanish E2E path
- OpenAPI determinism: skip when Node major < .nvmrc (matches frontend_npm.sh)
- chore(pre-commit): refresh .secrets.baseline (detect-secrets line drift)

Made-with: Cursor
