+++ .secrets.baseline
@@ -1516,7 +1516,7 @@
         "filename": "tests/test_web_interface_spanish.py",
         "hashed_secret": "00942f4668670f34c5943cf52c7ef3139fe2b8d6",
         "is_verified": false,
-        "line_number": 20
+        "line_number": 21
       }
     ],
     "tests/test_weekly_planning_blocks_97.py": [
@@ -1536,5 +1536,5 @@
       }
     ]
   },
-  "generated_at": "2026-03-22T13:03:15Z"
+  "generated_at": "2026-03-22T21:21:45Z"
 }
