+++ docs/deploy/SPA_APEX_ROUTING_CONTRACT.md
@@ -37,8 +37,15 @@ Caddy evaluates **POST**, then **OPTIONS**, then **GET** (legacy-only paths), th
 | `/ws*` | WebSocket foundation |
 | `/docs*`, `/redoc*`, `/openapi.json` | OpenAPI / docs |
 | `/admin*` | Admin |
-| `/privacy`, `/terms` | Legacy HTML (no client route today) |
+| `/privacy`, `/terms` | Legal JSON endpoints (no SPA route today) |
+| `/legacy*` | FastAPI-only legacy surfaces (embedded HTML BMI UI: `legacy_app.py:1494`) |
 | `/debug_env` | Debug (gate in prod env) |
+
+**Caddy matcher evidence:** `/legacy*` is included in the `@api` path list in [`deploy/Caddyfile.production:42`](../../deploy/Caddyfile.production).
+
+## Direct uvicorn / bypass Caddy
+
+When traffic hits **FastAPI only** (port `8000`, misconfigured clients, internal probes), **`GET /`** returns a **small JSON probe** (stable `service` / `surface` / `links`; handler `legacy_app.py:1486`, payload builder `app/bootstrap/direct_api_root.py:18`). The historical embedded HTML calculator is at **`GET /legacy/bmi-calculator`** (`legacy_app.py:1494`, template `app/bootstrap/legacy_bmi_web_html.py:9`). Production browsers still receive **`text/html`** for **`GET /`** from Caddy’s `file_server` at apex; they do not see this JSON unless they bypass the edge.
 
 ## Static (Caddy `file_server`)
 
@@ -56,7 +63,8 @@ Caddy evaluates **POST**, then **OPTIONS**, then **GET** (legacy-only paths), th
 
 ## QA smoke checklist (after deploy)
 
-- **MIME:** `GET /` returns `text/html`; `GET /health` returns JSON (via proxy).
+- **MIME (through Caddy):** `GET /` returns SPA `text/html` from static `file_server`; `GET /health` returns JSON (via proxy).
+- **Direct API:** `GET /` on uvicorn returns JSON probe (`app/bootstrap/direct_api_root.py:18`); legacy HTML UI: `GET /legacy/bmi-calculator` (proxied via `/legacy*` in `deploy/Caddyfile.production:42`).
 - **Deep link:** `GET /bmi` serves SPA `index.html` (not API); `GET /plan` (no SPA route) is proxied to the app.
 - **Legacy POST:** `POST /bmi` (and peers) reaches FastAPI (not static 405 from `file_server`).
 - **OpenAPI:** `GET /openapi.json` proxied (200, JSON).
