+++ tests/test_spanish_end_to_end_smoke.py
@@ -97,8 +97,8 @@ class TestSpanishEndToEndSmoke:
         # Verify Spanish language is processed (notes should be localized)
         assert isinstance(bmi_pro_result.get("notes"), list)
 
-        # 4. Test web interface with Spanish language parameter
-        web_response = self.client.get("/?lang=es")
+        # 4. Legacy embedded BMI web UI (Spanish) — canonical path (GET / is JSON probe)
+        web_response = self.client.get("/legacy/bmi-calculator?lang=es")
         assert web_response.status_code == 200
         assert "text/html" in web_response.headers["content-type"]
